% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generar_binomial.R, R/hello.R
\name{generar_binomial}
\alias{generar_binomial}
\title{Generar números binomiales aleatorios}
\usage{
generar_binomial(n, size, prob)

generar_binomial(n, size, prob)
}
\arguments{
\item{n}{Número de observaciones}

\item{size}{Número de ensayos por observación}

\item{prob}{Probabilidad de éxito en cada ensayo}
}
\value{
Vector con números binomiales generados

Vector con números binomiales generados
}
\description{
Esta función genera una muestra de números aleatorios desde una distribución binomial.

Esta función genera una muestra de números aleatorios desde una distribución binomial.
}
\examples{
generar_binomial(10, size = 5, prob = 0.3)
generar_binomial(10, size = 5, prob = 0.3)
}
