% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generar_normal.R
\name{generar_normal}
\alias{generar_normal}
\title{Genera números aleatorios de una distribución normal}
\usage{
generar_normal(n, mean = 0, sd = 1)
}
\arguments{
\item{n}{Número de valores a generar.}

\item{mean}{Media de la distribución normal. Valor predeterminado es 0.}

\item{sd}{Desviación estándar de la distribución normal. Valor predeterminado es 1.}
}
\value{
Un vector de números aleatorios de la distribución normal.
}
\description{
Esta función genera números aleatorios de una distribución normal con parámetros definidos por el usuario.
}
