% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generar_poisson.R
\name{generar_poisson}
\alias{generar_poisson}
\title{Genera números aleatorios de una distribución Poisson}
\usage{
generar_poisson(n, lambda = 1)
}
\arguments{
\item{n}{Número de valores a generar.}

\item{lambda}{Parámetro de tasa de la distribución de Poisson. Debe ser positivo.}
}
\value{
Un vector de números aleatorios de la distribución de Poisson.
}
\description{
Esta función genera números aleatorios de una distribución de Poisson con un parámetro lambda definido por el usuario.
}
\examples{
generar_poisson(10, lambda = 3)
}
